+++ packages/storage/src/typed.rs
@@ -68,12 +68,12 @@ where
         start: Option<&[u8]>,
         end: Option<&[u8]>,
         order: Order,
-    ) -> StdResult<Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b>> {
+    ) -> Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b> {
         let mapped = self
             .storage
-            .range(start, end, order)?
+            .range(start, end, order)
             .map(deserialize_kv::<T>);
-        Ok(Box::new(mapped))
+        Box::new(mapped)
     }
 
     /// update will load the data, perform the specified action, and store the result
@@ -130,12 +130,12 @@ where
         start: Option<&[u8]>,
         end: Option<&[u8]>,
         order: Order,
-    ) -> StdResult<Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b>> {
+    ) -> Box<dyn Iterator<Item = StdResult<KV<T>>> + 'b> {
         let mapped = self
             .storage
-            .range(start, end, order)?
+            .range(start, end, order)
             .map(deserialize_kv::<T>);
-        Ok(Box::new(mapped))
+        Box::new(mapped)
     }
 }
 
@@ -309,10 +309,8 @@ mod test {
         bucket.save(b"maria", &maria).unwrap();
         bucket.save(b"jose", &jose).unwrap();
 
-        let res_data: StdResult<Vec<KV<Data>>> = bucket
-            .range(None, None, Order::Ascending)
-            .unwrap()
-            .collect();
+        let res_data: StdResult<Vec<KV<Data>>> =
+            bucket.range(None, None, Order::Ascending).collect();
         let data = res_data.unwrap();
         assert_eq!(data.len(), 2);
         assert_eq!(data[0], (b"jose".to_vec(), jose.clone()));
@@ -320,10 +318,8 @@ mod test {
 
         // also works for readonly
         let read_bucket = typed_read::<_, Data>(&store);
-        let res_data: StdResult<Vec<KV<Data>>> = read_bucket
-            .range(None, None, Order::Ascending)
-            .unwrap()
-            .collect();
+        let res_data: StdResult<Vec<KV<Data>>> =
+            read_bucket.range(None, None, Order::Ascending).collect();
         let data = res_data.unwrap();
         assert_eq!(data.len(), 2);
         assert_eq!(data[0], (b"jose".to_vec(), jose));